# Improved performance by using a circular dynamically sized dup table