Inspired by #9 but less intrusive changes. Since it scales more aggressively there is a chance it uses more memory, but at the same time it should have less overhead per dup entry on average owing to its flat format.

The main benefit of this is that the performance becomes more predictable with less tuning required. It also allocates fewer objects so should reduce GC pressure significantly.

## Before change
| Method | DataSet         | Pool                 | Mean        | Error     | StdDev     | Median      |
|------- |---------------- |--------------------- |------------:|----------:|-----------:|------------:|
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 4,639.18 ns | 92.503 ns | 266.894 ns | 4,563.86 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   401.31 ns |  8.005 ns |   7.862 ns |   399.50 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |   115.02 ns |  0.352 ns |   0.312 ns |   114.98 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 1,533.70 ns | 18.462 ns |  16.366 ns | 1,536.82 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   209.45 ns |  5.817 ns |  16.782 ns |   205.75 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    90.47 ns |  1.696 ns |   1.503 ns |    90.31 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] |   645.70 ns | 12.502 ns |  14.882 ns |   643.25 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   139.66 ns |  2.829 ns |   7.403 ns |   137.83 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    61.68 ns |  0.890 ns |   0.789 ns |    61.68 ns |

## After change
| Method | DataSet         | Pool                 | Mean      | Error    | StdDev    |
|------- |---------------- |--------------------- |----------:|---------:|----------:|
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 437.56 ns | 8.702 ns | 15.912 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] | 105.96 ns | 2.142 ns |  3.917 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |  74.33 ns | 0.998 ns |  0.833 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 410.85 ns | 7.743 ns |  7.243 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |  99.77 ns | 2.003 ns |  1.776 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |  74.41 ns | 1.467 ns |  1.225 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 431.01 ns | 8.385 ns | 10.298 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] | 111.95 ns | 4.492 ns | 12.743 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |  59.24 ns | 0.826 ns |  0.773 ns |